Convolution kernels

Creates a convolution kernel

uniformKernel(r = NULL, mu = NULL, cellDim = 1)

exponentialKernel(mu, cellDim = 1, negligible = 10^-10)

Arguments

r

The radius of the kernel, in physical units.

mu

Alternative to r, so that r = 3 * mu / 2 (equation for the average distance from the center of a circle).

cellDim

Default to 1, cell dimensions used as a scaling factor for the kernel size

negligible

Negligible value when rounding edges of exponential kernels.

Value

A kernel (matrix).

Examples

uniformKernel(r = 1, mu = 2)
#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,] 0.00000000 0.00000000 0.00000000 0.03448276 0.00000000 0.00000000
#> [2,] 0.00000000 0.03448276 0.03448276 0.03448276 0.03448276 0.03448276
#> [3,] 0.00000000 0.03448276 0.03448276 0.03448276 0.03448276 0.03448276
#> [4,] 0.03448276 0.03448276 0.03448276 0.03448276 0.03448276 0.03448276
#> [5,] 0.00000000 0.03448276 0.03448276 0.03448276 0.03448276 0.03448276
#> [6,] 0.00000000 0.03448276 0.03448276 0.03448276 0.03448276 0.03448276
#> [7,] 0.00000000 0.00000000 0.00000000 0.03448276 0.00000000 0.00000000
#>            [,7]
#> [1,] 0.00000000
#> [2,] 0.00000000
#> [3,] 0.00000000
#> [4,] 0.03448276
#> [5,] 0.00000000
#> [6,] 0.00000000
#> [7,] 0.00000000
exponentialKernel(mu = 2)
#>               [,1]         [,2]         [,3]         [,4]         [,5]
#>  [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [2,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [3,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [4,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [5,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [6,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [7,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [8,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 1.711563e-10
#>  [9,] 0.000000e+00 0.000000e+00 0.000000e+00 1.410978e-10 3.167477e-10
#> [10,] 0.000000e+00 0.000000e+00 1.059622e-10 2.470643e-10 5.677139e-10
#> [11,] 0.000000e+00 0.000000e+00 1.753607e-10 4.178082e-10 9.825421e-10
#> [12,] 0.000000e+00 1.137849e-10 2.796374e-10 6.804276e-10 1.636877e-09
#> [13,] 0.000000e+00 1.711563e-10 4.285425e-10 1.064022e-09 2.616418e-09
#> [14,] 0.000000e+00 2.470643e-10 6.294822e-10 1.592928e-09 3.999227e-09
#> [15,] 1.313010e-10 3.414655e-10 8.839845e-10 2.276397e-09 5.826070e-09
#> [16,] 1.711563e-10 4.508897e-10 1.183893e-09 3.096601e-09 8.063080e-09
#> [17,] 2.131567e-10 5.677139e-10 1.508719e-09 3.999227e-09 1.056903e-08
#> [18,] 2.532449e-10 6.804276e-10 1.825881e-09 4.892355e-09 1.308583e-08
#> [19,] 2.866764e-10 7.752087e-10 2.095055e-09 5.658209e-09 1.526915e-08
#> [20,] 3.089320e-10 8.386621e-10 2.276397e-09 6.177810e-09 1.676227e-08
#> [21,] 3.167477e-10 8.610094e-10 2.340466e-09 6.362047e-09 1.729384e-08
#> [22,] 3.089320e-10 8.386621e-10 2.276397e-09 6.177810e-09 1.676227e-08
#> [23,] 2.866764e-10 7.752087e-10 2.095055e-09 5.658209e-09 1.526915e-08
#> [24,] 2.532449e-10 6.804276e-10 1.825881e-09 4.892355e-09 1.308583e-08
#> [25,] 2.131567e-10 5.677139e-10 1.508719e-09 3.999227e-09 1.056903e-08
#> [26,] 1.711563e-10 4.508897e-10 1.183893e-09 3.096601e-09 8.063080e-09
#> [27,] 1.313010e-10 3.414655e-10 8.839845e-10 2.276397e-09 5.826070e-09
#> [28,] 0.000000e+00 2.470643e-10 6.294822e-10 1.592928e-09 3.999227e-09
#> [29,] 0.000000e+00 1.711563e-10 4.285425e-10 1.064022e-09 2.616418e-09
#> [30,] 0.000000e+00 1.137849e-10 2.796374e-10 6.804276e-10 1.636877e-09
#> [31,] 0.000000e+00 0.000000e+00 1.753607e-10 4.178082e-10 9.825421e-10
#> [32,] 0.000000e+00 0.000000e+00 1.059622e-10 2.470643e-10 5.677139e-10
#> [33,] 0.000000e+00 0.000000e+00 0.000000e+00 1.410978e-10 3.167477e-10
#> [34,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 1.711563e-10
#> [35,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [36,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [37,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [38,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [39,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [40,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [41,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>               [,6]         [,7]         [,8]         [,9]        [,10]
#>  [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [2,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [3,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 1.059622e-10
#>  [4,] 0.000000e+00 0.000000e+00 0.000000e+00 1.410978e-10 2.470643e-10
#>  [5,] 0.000000e+00 0.000000e+00 1.711563e-10 3.167477e-10 5.677139e-10
#>  [6,] 0.000000e+00 1.886363e-10 3.682169e-10 6.983586e-10 1.283095e-09
#>  [7,] 1.886363e-10 3.872675e-10 7.752087e-10 1.508719e-09 2.845835e-09
#>  [8,] 3.682169e-10 7.752087e-10 1.592928e-09 3.185293e-09 6.177810e-09
#>  [9,] 6.983586e-10 1.508719e-09 3.185293e-09 6.552112e-09 1.308583e-08
#> [10,] 1.283095e-09 2.845835e-09 6.177810e-09 1.308583e-08 2.695048e-08
#> [11,] 2.276397e-09 5.184636e-09 1.157828e-08 2.527589e-08 5.374666e-08
#> [12,] 3.886444e-09 9.088712e-09 2.088355e-08 4.700952e-08 1.033044e-07
#> [13,] 6.362047e-09 1.526915e-08 3.608989e-08 8.377588e-08 1.903529e-07
#> [14,] 9.948331e-09 2.448044e-08 5.947328e-08 1.422955e-07 3.342688e-07
#> [15,] 1.480338e-08 3.729294e-08 9.299346e-08 2.290505e-07 5.557875e-07
#> [16,] 2.088355e-08 5.374666e-08 1.372717e-07 3.473560e-07 8.689662e-07
#> [17,] 2.783170e-08 7.297598e-08 1.903529e-07 4.933602e-07 1.268576e-06
#> [18,] 3.492810e-08 9.299346e-08 2.468266e-07 6.526450e-07 1.717450e-06
#> [19,] 4.116565e-08 1.108541e-07 2.980925e-07 8.001707e-07 2.143104e-06
#> [20,] 4.547005e-08 1.233075e-07 3.342688e-07 9.057430e-07 2.452809e-06
#> [21,] 4.700952e-08 1.277851e-07 3.473560e-07 9.442115e-07 2.566633e-06
#> [22,] 4.547005e-08 1.233075e-07 3.342688e-07 9.057430e-07 2.452809e-06
#> [23,] 4.116565e-08 1.108541e-07 2.980925e-07 8.001707e-07 2.143104e-06
#> [24,] 3.492810e-08 9.299346e-08 2.468266e-07 6.526450e-07 1.717450e-06
#> [25,] 2.783170e-08 7.297598e-08 1.903529e-07 4.933602e-07 1.268576e-06
#> [26,] 2.088355e-08 5.374666e-08 1.372717e-07 3.473560e-07 8.689662e-07
#> [27,] 1.480338e-08 3.729294e-08 9.299346e-08 2.290505e-07 5.557875e-07
#> [28,] 9.948331e-09 2.448044e-08 5.947328e-08 1.422955e-07 3.342688e-07
#> [29,] 6.362047e-09 1.526915e-08 3.608989e-08 8.377588e-08 1.903529e-07
#> [30,] 3.886444e-09 9.088712e-09 2.088355e-08 4.700952e-08 1.033044e-07
#> [31,] 2.276397e-09 5.184636e-09 1.157828e-08 2.527589e-08 5.374666e-08
#> [32,] 1.283095e-09 2.845835e-09 6.177810e-09 1.308583e-08 2.695048e-08
#> [33,] 6.983586e-10 1.508719e-09 3.185293e-09 6.552112e-09 1.308583e-08
#> [34,] 3.682169e-10 7.752087e-10 1.592928e-09 3.185293e-09 6.177810e-09
#> [35,] 1.886363e-10 3.872675e-10 7.752087e-10 1.508719e-09 2.845835e-09
#> [36,] 0.000000e+00 1.886363e-10 3.682169e-10 6.983586e-10 1.283095e-09
#> [37,] 0.000000e+00 0.000000e+00 1.711563e-10 3.167477e-10 5.677139e-10
#> [38,] 0.000000e+00 0.000000e+00 0.000000e+00 1.410978e-10 2.470643e-10
#> [39,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 1.059622e-10
#> [40,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [41,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>              [,11]        [,12]        [,13]        [,14]        [,15]
#>  [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 1.313010e-10
#>  [2,] 0.000000e+00 1.137849e-10 1.711563e-10 2.470643e-10 3.414655e-10
#>  [3,] 1.753607e-10 2.796374e-10 4.285425e-10 6.294822e-10 8.839845e-10
#>  [4,] 4.178082e-10 6.804276e-10 1.064022e-09 1.592928e-09 2.276397e-09
#>  [5,] 9.825421e-10 1.636877e-09 2.616418e-09 3.999227e-09 5.826070e-09
#>  [6,] 2.276397e-09 3.886444e-09 6.362047e-09 9.948331e-09 1.480338e-08
#>  [7,] 5.184636e-09 9.088712e-09 1.526915e-08 2.448044e-08 3.729294e-08
#>  [8,] 1.157828e-08 2.088355e-08 3.608989e-08 5.947328e-08 9.299346e-08
#>  [9,] 2.527589e-08 4.700952e-08 8.377588e-08 1.422955e-07 2.290505e-07
#> [10,] 5.374666e-08 1.033044e-07 1.903529e-07 3.342688e-07 5.557875e-07
#> [11,] 1.108541e-07 2.206827e-07 4.216189e-07 7.680037e-07 1.324045e-06
#> [12,] 2.206827e-07 4.559705e-07 9.057430e-07 1.717450e-06 3.083117e-06
#> [13,] 4.216189e-07 9.057430e-07 1.875521e-06 3.715262e-06 6.976832e-06
#> [14,] 7.680037e-07 1.717450e-06 3.715262e-06 7.714487e-06 1.522670e-05
#> [15,] 1.324045e-06 3.083117e-06 6.976832e-06 1.522670e-05 3.173162e-05
#> [16,] 2.143104e-06 5.191199e-06 1.228788e-05 2.822674e-05 6.232392e-05
#> [17,] 3.229300e-06 8.115186e-06 2.005188e-05 4.844054e-05 1.134648e-04
#> [18,] 4.491957e-06 1.165531e-05 2.992192e-05 7.570330e-05 1.876148e-04
#> [19,] 5.723359e-06 1.522670e-05 4.030130e-05 1.058989e-04 2.753491e-04
#> [20,] 6.637393e-06 1.794317e-05 4.844054e-05 1.305201e-04 3.506654e-04
#> [21,] 6.976832e-06 1.896500e-05 5.155220e-05 1.401334e-04 3.809221e-04
#> [22,] 6.637393e-06 1.794317e-05 4.844054e-05 1.305201e-04 3.506654e-04
#> [23,] 5.723359e-06 1.522670e-05 4.030130e-05 1.058989e-04 2.753491e-04
#> [24,] 4.491957e-06 1.165531e-05 2.992192e-05 7.570330e-05 1.876148e-04
#> [25,] 3.229300e-06 8.115186e-06 2.005188e-05 4.844054e-05 1.134648e-04
#> [26,] 2.143104e-06 5.191199e-06 1.228788e-05 2.822674e-05 6.232392e-05
#> [27,] 1.324045e-06 3.083117e-06 6.976832e-06 1.522670e-05 3.173162e-05
#> [28,] 7.680037e-07 1.717450e-06 3.715262e-06 7.714487e-06 1.522670e-05
#> [29,] 4.216189e-07 9.057430e-07 1.875521e-06 3.715262e-06 6.976832e-06
#> [30,] 2.206827e-07 4.559705e-07 9.057430e-07 1.717450e-06 3.083117e-06
#> [31,] 1.108541e-07 2.206827e-07 4.216189e-07 7.680037e-07 1.324045e-06
#> [32,] 5.374666e-08 1.033044e-07 1.903529e-07 3.342688e-07 5.557875e-07
#> [33,] 2.527589e-08 4.700952e-08 8.377588e-08 1.422955e-07 2.290505e-07
#> [34,] 1.157828e-08 2.088355e-08 3.608989e-08 5.947328e-08 9.299346e-08
#> [35,] 5.184636e-09 9.088712e-09 1.526915e-08 2.448044e-08 3.729294e-08
#> [36,] 2.276397e-09 3.886444e-09 6.362047e-09 9.948331e-09 1.480338e-08
#> [37,] 9.825421e-10 1.636877e-09 2.616418e-09 3.999227e-09 5.826070e-09
#> [38,] 4.178082e-10 6.804276e-10 1.064022e-09 1.592928e-09 2.276397e-09
#> [39,] 1.753607e-10 2.796374e-10 4.285425e-10 6.294822e-10 8.839845e-10
#> [40,] 0.000000e+00 1.137849e-10 1.711563e-10 2.470643e-10 3.414655e-10
#> [41,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 1.313010e-10
#>              [,16]        [,17]        [,18]        [,19]        [,20]
#>  [1,] 1.711563e-10 2.131567e-10 2.532449e-10 2.866764e-10 3.089320e-10
#>  [2,] 4.508897e-10 5.677139e-10 6.804276e-10 7.752087e-10 8.386621e-10
#>  [3,] 1.183893e-09 1.508719e-09 1.825881e-09 2.095055e-09 2.276397e-09
#>  [4,] 3.096601e-09 3.999227e-09 4.892355e-09 5.658209e-09 6.177810e-09
#>  [5,] 8.063080e-09 1.056903e-08 1.308583e-08 1.526915e-08 1.676227e-08
#>  [6,] 2.088355e-08 2.783170e-08 3.492810e-08 4.116565e-08 4.547005e-08
#>  [7,] 5.374666e-08 7.297598e-08 9.299346e-08 1.108541e-07 1.233075e-07
#>  [8,] 1.372717e-07 1.903529e-07 2.468266e-07 2.980925e-07 3.342688e-07
#>  [9,] 3.473560e-07 4.933602e-07 6.526450e-07 8.001707e-07 9.057430e-07
#> [10,] 8.689662e-07 1.268576e-06 1.717450e-06 2.143104e-06 2.452809e-06
#> [11,] 2.143104e-06 3.229300e-06 4.491957e-06 5.723359e-06 6.637393e-06
#> [12,] 5.191199e-06 8.115186e-06 1.165531e-05 1.522670e-05 1.794317e-05
#> [13,] 1.228788e-05 2.005188e-05 2.992192e-05 4.030130e-05 4.844054e-05
#> [14,] 2.822674e-05 4.844054e-05 7.570330e-05 1.058989e-04 1.305201e-04
#> [15,] 6.232392e-05 1.134648e-04 1.876148e-04 2.753491e-04 3.506654e-04
#> [16,] 1.305201e-04 2.545432e-04 4.510792e-04 7.044590e-04 9.378363e-04
#> [17,] 2.545432e-04 5.368618e-04 1.035454e-03 1.755412e-03 2.488634e-03
#> [18,] 4.510792e-04 1.035454e-03 2.208247e-03 4.175727e-03 6.504941e-03
#> [19,] 7.044590e-04 1.755412e-03 4.175727e-03 9.083072e-03 1.642446e-02
#> [20,] 9.378363e-04 2.488634e-03 6.504941e-03 1.642446e-02 3.736095e-02
#> [21,] 1.035454e-03 2.814655e-03 7.651025e-03 2.079764e-02 5.653385e-02
#> [22,] 9.378363e-04 2.488634e-03 6.504941e-03 1.642446e-02 3.736095e-02
#> [23,] 7.044590e-04 1.755412e-03 4.175727e-03 9.083072e-03 1.642446e-02
#> [24,] 4.510792e-04 1.035454e-03 2.208247e-03 4.175727e-03 6.504941e-03
#> [25,] 2.545432e-04 5.368618e-04 1.035454e-03 1.755412e-03 2.488634e-03
#> [26,] 1.305201e-04 2.545432e-04 4.510792e-04 7.044590e-04 9.378363e-04
#> [27,] 6.232392e-05 1.134648e-04 1.876148e-04 2.753491e-04 3.506654e-04
#> [28,] 2.822674e-05 4.844054e-05 7.570330e-05 1.058989e-04 1.305201e-04
#> [29,] 1.228788e-05 2.005188e-05 2.992192e-05 4.030130e-05 4.844054e-05
#> [30,] 5.191199e-06 8.115186e-06 1.165531e-05 1.522670e-05 1.794317e-05
#> [31,] 2.143104e-06 3.229300e-06 4.491957e-06 5.723359e-06 6.637393e-06
#> [32,] 8.689662e-07 1.268576e-06 1.717450e-06 2.143104e-06 2.452809e-06
#> [33,] 3.473560e-07 4.933602e-07 6.526450e-07 8.001707e-07 9.057430e-07
#> [34,] 1.372717e-07 1.903529e-07 2.468266e-07 2.980925e-07 3.342688e-07
#> [35,] 5.374666e-08 7.297598e-08 9.299346e-08 1.108541e-07 1.233075e-07
#> [36,] 2.088355e-08 2.783170e-08 3.492810e-08 4.116565e-08 4.547005e-08
#> [37,] 8.063080e-09 1.056903e-08 1.308583e-08 1.526915e-08 1.676227e-08
#> [38,] 3.096601e-09 3.999227e-09 4.892355e-09 5.658209e-09 6.177810e-09
#> [39,] 1.183893e-09 1.508719e-09 1.825881e-09 2.095055e-09 2.276397e-09
#> [40,] 4.508897e-10 5.677139e-10 6.804276e-10 7.752087e-10 8.386621e-10
#> [41,] 1.711563e-10 2.131567e-10 2.532449e-10 2.866764e-10 3.089320e-10
#>              [,21]        [,22]        [,23]        [,24]        [,25]
#>  [1,] 3.167477e-10 3.089320e-10 2.866764e-10 2.532449e-10 2.131567e-10
#>  [2,] 8.610094e-10 8.386621e-10 7.752087e-10 6.804276e-10 5.677139e-10
#>  [3,] 2.340466e-09 2.276397e-09 2.095055e-09 1.825881e-09 1.508719e-09
#>  [4,] 6.362047e-09 6.177810e-09 5.658209e-09 4.892355e-09 3.999227e-09
#>  [5,] 1.729384e-08 1.676227e-08 1.526915e-08 1.308583e-08 1.056903e-08
#>  [6,] 4.700952e-08 4.547005e-08 4.116565e-08 3.492810e-08 2.783170e-08
#>  [7,] 1.277851e-07 1.233075e-07 1.108541e-07 9.299346e-08 7.297598e-08
#>  [8,] 3.473560e-07 3.342688e-07 2.980925e-07 2.468266e-07 1.903529e-07
#>  [9,] 9.442115e-07 9.057430e-07 8.001707e-07 6.526450e-07 4.933602e-07
#> [10,] 2.566633e-06 2.452809e-06 2.143104e-06 1.717450e-06 1.268576e-06
#> [11,] 6.976832e-06 6.637393e-06 5.723359e-06 4.491957e-06 3.229300e-06
#> [12,] 1.896500e-05 1.794317e-05 1.522670e-05 1.165531e-05 8.115186e-06
#> [13,] 5.155220e-05 4.844054e-05 4.030130e-05 2.992192e-05 2.005188e-05
#> [14,] 1.401334e-04 1.305201e-04 1.058989e-04 7.570330e-05 4.844054e-05
#> [15,] 3.809221e-04 3.506654e-04 2.753491e-04 1.876148e-04 1.134648e-04
#> [16,] 1.035454e-03 9.378363e-04 7.044590e-04 4.510792e-04 2.545432e-04
#> [17,] 2.814655e-03 2.488634e-03 1.755412e-03 1.035454e-03 5.368618e-04
#> [18,] 7.651025e-03 6.504941e-03 4.175727e-03 2.208247e-03 1.035454e-03
#> [19,] 2.079764e-02 1.642446e-02 9.083072e-03 4.175727e-03 1.755412e-03
#> [20,] 5.653385e-02 3.736095e-02 1.642446e-02 6.504941e-03 2.488634e-03
#> [21,] 1.536749e-01 5.653385e-02 2.079764e-02 7.651025e-03 2.814655e-03
#> [22,] 5.653385e-02 3.736095e-02 1.642446e-02 6.504941e-03 2.488634e-03
#> [23,] 2.079764e-02 1.642446e-02 9.083072e-03 4.175727e-03 1.755412e-03
#> [24,] 7.651025e-03 6.504941e-03 4.175727e-03 2.208247e-03 1.035454e-03
#> [25,] 2.814655e-03 2.488634e-03 1.755412e-03 1.035454e-03 5.368618e-04
#> [26,] 1.035454e-03 9.378363e-04 7.044590e-04 4.510792e-04 2.545432e-04
#> [27,] 3.809221e-04 3.506654e-04 2.753491e-04 1.876148e-04 1.134648e-04
#> [28,] 1.401334e-04 1.305201e-04 1.058989e-04 7.570330e-05 4.844054e-05
#> [29,] 5.155220e-05 4.844054e-05 4.030130e-05 2.992192e-05 2.005188e-05
#> [30,] 1.896500e-05 1.794317e-05 1.522670e-05 1.165531e-05 8.115186e-06
#> [31,] 6.976832e-06 6.637393e-06 5.723359e-06 4.491957e-06 3.229300e-06
#> [32,] 2.566633e-06 2.452809e-06 2.143104e-06 1.717450e-06 1.268576e-06
#> [33,] 9.442115e-07 9.057430e-07 8.001707e-07 6.526450e-07 4.933602e-07
#> [34,] 3.473560e-07 3.342688e-07 2.980925e-07 2.468266e-07 1.903529e-07
#> [35,] 1.277851e-07 1.233075e-07 1.108541e-07 9.299346e-08 7.297598e-08
#> [36,] 4.700952e-08 4.547005e-08 4.116565e-08 3.492810e-08 2.783170e-08
#> [37,] 1.729384e-08 1.676227e-08 1.526915e-08 1.308583e-08 1.056903e-08
#> [38,] 6.362047e-09 6.177810e-09 5.658209e-09 4.892355e-09 3.999227e-09
#> [39,] 2.340466e-09 2.276397e-09 2.095055e-09 1.825881e-09 1.508719e-09
#> [40,] 8.610094e-10 8.386621e-10 7.752087e-10 6.804276e-10 5.677139e-10
#> [41,] 3.167477e-10 3.089320e-10 2.866764e-10 2.532449e-10 2.131567e-10
#>              [,26]        [,27]        [,28]        [,29]        [,30]
#>  [1,] 1.711563e-10 1.313010e-10 0.000000e+00 0.000000e+00 0.000000e+00
#>  [2,] 4.508897e-10 3.414655e-10 2.470643e-10 1.711563e-10 1.137849e-10
#>  [3,] 1.183893e-09 8.839845e-10 6.294822e-10 4.285425e-10 2.796374e-10
#>  [4,] 3.096601e-09 2.276397e-09 1.592928e-09 1.064022e-09 6.804276e-10
#>  [5,] 8.063080e-09 5.826070e-09 3.999227e-09 2.616418e-09 1.636877e-09
#>  [6,] 2.088355e-08 1.480338e-08 9.948331e-09 6.362047e-09 3.886444e-09
#>  [7,] 5.374666e-08 3.729294e-08 2.448044e-08 1.526915e-08 9.088712e-09
#>  [8,] 1.372717e-07 9.299346e-08 5.947328e-08 3.608989e-08 2.088355e-08
#>  [9,] 3.473560e-07 2.290505e-07 1.422955e-07 8.377588e-08 4.700952e-08
#> [10,] 8.689662e-07 5.557875e-07 3.342688e-07 1.903529e-07 1.033044e-07
#> [11,] 2.143104e-06 1.324045e-06 7.680037e-07 4.216189e-07 2.206827e-07
#> [12,] 5.191199e-06 3.083117e-06 1.717450e-06 9.057430e-07 4.559705e-07
#> [13,] 1.228788e-05 6.976832e-06 3.715262e-06 1.875521e-06 9.057430e-07
#> [14,] 2.822674e-05 1.522670e-05 7.714487e-06 3.715262e-06 1.717450e-06
#> [15,] 6.232392e-05 3.173162e-05 1.522670e-05 6.976832e-06 3.083117e-06
#> [16,] 1.305201e-04 6.232392e-05 2.822674e-05 1.228788e-05 5.191199e-06
#> [17,] 2.545432e-04 1.134648e-04 4.844054e-05 2.005188e-05 8.115186e-06
#> [18,] 4.510792e-04 1.876148e-04 7.570330e-05 2.992192e-05 1.165531e-05
#> [19,] 7.044590e-04 2.753491e-04 1.058989e-04 4.030130e-05 1.522670e-05
#> [20,] 9.378363e-04 3.506654e-04 1.305201e-04 4.844054e-05 1.794317e-05
#> [21,] 1.035454e-03 3.809221e-04 1.401334e-04 5.155220e-05 1.896500e-05
#> [22,] 9.378363e-04 3.506654e-04 1.305201e-04 4.844054e-05 1.794317e-05
#> [23,] 7.044590e-04 2.753491e-04 1.058989e-04 4.030130e-05 1.522670e-05
#> [24,] 4.510792e-04 1.876148e-04 7.570330e-05 2.992192e-05 1.165531e-05
#> [25,] 2.545432e-04 1.134648e-04 4.844054e-05 2.005188e-05 8.115186e-06
#> [26,] 1.305201e-04 6.232392e-05 2.822674e-05 1.228788e-05 5.191199e-06
#> [27,] 6.232392e-05 3.173162e-05 1.522670e-05 6.976832e-06 3.083117e-06
#> [28,] 2.822674e-05 1.522670e-05 7.714487e-06 3.715262e-06 1.717450e-06
#> [29,] 1.228788e-05 6.976832e-06 3.715262e-06 1.875521e-06 9.057430e-07
#> [30,] 5.191199e-06 3.083117e-06 1.717450e-06 9.057430e-07 4.559705e-07
#> [31,] 2.143104e-06 1.324045e-06 7.680037e-07 4.216189e-07 2.206827e-07
#> [32,] 8.689662e-07 5.557875e-07 3.342688e-07 1.903529e-07 1.033044e-07
#> [33,] 3.473560e-07 2.290505e-07 1.422955e-07 8.377588e-08 4.700952e-08
#> [34,] 1.372717e-07 9.299346e-08 5.947328e-08 3.608989e-08 2.088355e-08
#> [35,] 5.374666e-08 3.729294e-08 2.448044e-08 1.526915e-08 9.088712e-09
#> [36,] 2.088355e-08 1.480338e-08 9.948331e-09 6.362047e-09 3.886444e-09
#> [37,] 8.063080e-09 5.826070e-09 3.999227e-09 2.616418e-09 1.636877e-09
#> [38,] 3.096601e-09 2.276397e-09 1.592928e-09 1.064022e-09 6.804276e-10
#> [39,] 1.183893e-09 8.839845e-10 6.294822e-10 4.285425e-10 2.796374e-10
#> [40,] 4.508897e-10 3.414655e-10 2.470643e-10 1.711563e-10 1.137849e-10
#> [41,] 1.711563e-10 1.313010e-10 0.000000e+00 0.000000e+00 0.000000e+00
#>              [,31]        [,32]        [,33]        [,34]        [,35]
#>  [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [2,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [3,] 1.753607e-10 1.059622e-10 0.000000e+00 0.000000e+00 0.000000e+00
#>  [4,] 4.178082e-10 2.470643e-10 1.410978e-10 0.000000e+00 0.000000e+00
#>  [5,] 9.825421e-10 5.677139e-10 3.167477e-10 1.711563e-10 0.000000e+00
#>  [6,] 2.276397e-09 1.283095e-09 6.983586e-10 3.682169e-10 1.886363e-10
#>  [7,] 5.184636e-09 2.845835e-09 1.508719e-09 7.752087e-10 3.872675e-10
#>  [8,] 1.157828e-08 6.177810e-09 3.185293e-09 1.592928e-09 7.752087e-10
#>  [9,] 2.527589e-08 1.308583e-08 6.552112e-09 3.185293e-09 1.508719e-09
#> [10,] 5.374666e-08 2.695048e-08 1.308583e-08 6.177810e-09 2.845835e-09
#> [11,] 1.108541e-07 5.374666e-08 2.527589e-08 1.157828e-08 5.184636e-09
#> [12,] 2.206827e-07 1.033044e-07 4.700952e-08 2.088355e-08 9.088712e-09
#> [13,] 4.216189e-07 1.903529e-07 8.377588e-08 3.608989e-08 1.526915e-08
#> [14,] 7.680037e-07 3.342688e-07 1.422955e-07 5.947328e-08 2.448044e-08
#> [15,] 1.324045e-06 5.557875e-07 2.290505e-07 9.299346e-08 3.729294e-08
#> [16,] 2.143104e-06 8.689662e-07 3.473560e-07 1.372717e-07 5.374666e-08
#> [17,] 3.229300e-06 1.268576e-06 4.933602e-07 1.903529e-07 7.297598e-08
#> [18,] 4.491957e-06 1.717450e-06 6.526450e-07 2.468266e-07 9.299346e-08
#> [19,] 5.723359e-06 2.143104e-06 8.001707e-07 2.980925e-07 1.108541e-07
#> [20,] 6.637393e-06 2.452809e-06 9.057430e-07 3.342688e-07 1.233075e-07
#> [21,] 6.976832e-06 2.566633e-06 9.442115e-07 3.473560e-07 1.277851e-07
#> [22,] 6.637393e-06 2.452809e-06 9.057430e-07 3.342688e-07 1.233075e-07
#> [23,] 5.723359e-06 2.143104e-06 8.001707e-07 2.980925e-07 1.108541e-07
#> [24,] 4.491957e-06 1.717450e-06 6.526450e-07 2.468266e-07 9.299346e-08
#> [25,] 3.229300e-06 1.268576e-06 4.933602e-07 1.903529e-07 7.297598e-08
#> [26,] 2.143104e-06 8.689662e-07 3.473560e-07 1.372717e-07 5.374666e-08
#> [27,] 1.324045e-06 5.557875e-07 2.290505e-07 9.299346e-08 3.729294e-08
#> [28,] 7.680037e-07 3.342688e-07 1.422955e-07 5.947328e-08 2.448044e-08
#> [29,] 4.216189e-07 1.903529e-07 8.377588e-08 3.608989e-08 1.526915e-08
#> [30,] 2.206827e-07 1.033044e-07 4.700952e-08 2.088355e-08 9.088712e-09
#> [31,] 1.108541e-07 5.374666e-08 2.527589e-08 1.157828e-08 5.184636e-09
#> [32,] 5.374666e-08 2.695048e-08 1.308583e-08 6.177810e-09 2.845835e-09
#> [33,] 2.527589e-08 1.308583e-08 6.552112e-09 3.185293e-09 1.508719e-09
#> [34,] 1.157828e-08 6.177810e-09 3.185293e-09 1.592928e-09 7.752087e-10
#> [35,] 5.184636e-09 2.845835e-09 1.508719e-09 7.752087e-10 3.872675e-10
#> [36,] 2.276397e-09 1.283095e-09 6.983586e-10 3.682169e-10 1.886363e-10
#> [37,] 9.825421e-10 5.677139e-10 3.167477e-10 1.711563e-10 0.000000e+00
#> [38,] 4.178082e-10 2.470643e-10 1.410978e-10 0.000000e+00 0.000000e+00
#> [39,] 1.753607e-10 1.059622e-10 0.000000e+00 0.000000e+00 0.000000e+00
#> [40,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [41,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>              [,36]        [,37]        [,38]        [,39]        [,40]
#>  [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [2,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [3,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [4,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [5,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [6,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [7,] 1.886363e-10 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [8,] 3.682169e-10 1.711563e-10 0.000000e+00 0.000000e+00 0.000000e+00
#>  [9,] 6.983586e-10 3.167477e-10 1.410978e-10 0.000000e+00 0.000000e+00
#> [10,] 1.283095e-09 5.677139e-10 2.470643e-10 1.059622e-10 0.000000e+00
#> [11,] 2.276397e-09 9.825421e-10 4.178082e-10 1.753607e-10 0.000000e+00
#> [12,] 3.886444e-09 1.636877e-09 6.804276e-10 2.796374e-10 1.137849e-10
#> [13,] 6.362047e-09 2.616418e-09 1.064022e-09 4.285425e-10 1.711563e-10
#> [14,] 9.948331e-09 3.999227e-09 1.592928e-09 6.294822e-10 2.470643e-10
#> [15,] 1.480338e-08 5.826070e-09 2.276397e-09 8.839845e-10 3.414655e-10
#> [16,] 2.088355e-08 8.063080e-09 3.096601e-09 1.183893e-09 4.508897e-10
#> [17,] 2.783170e-08 1.056903e-08 3.999227e-09 1.508719e-09 5.677139e-10
#> [18,] 3.492810e-08 1.308583e-08 4.892355e-09 1.825881e-09 6.804276e-10
#> [19,] 4.116565e-08 1.526915e-08 5.658209e-09 2.095055e-09 7.752087e-10
#> [20,] 4.547005e-08 1.676227e-08 6.177810e-09 2.276397e-09 8.386621e-10
#> [21,] 4.700952e-08 1.729384e-08 6.362047e-09 2.340466e-09 8.610094e-10
#> [22,] 4.547005e-08 1.676227e-08 6.177810e-09 2.276397e-09 8.386621e-10
#> [23,] 4.116565e-08 1.526915e-08 5.658209e-09 2.095055e-09 7.752087e-10
#> [24,] 3.492810e-08 1.308583e-08 4.892355e-09 1.825881e-09 6.804276e-10
#> [25,] 2.783170e-08 1.056903e-08 3.999227e-09 1.508719e-09 5.677139e-10
#> [26,] 2.088355e-08 8.063080e-09 3.096601e-09 1.183893e-09 4.508897e-10
#> [27,] 1.480338e-08 5.826070e-09 2.276397e-09 8.839845e-10 3.414655e-10
#> [28,] 9.948331e-09 3.999227e-09 1.592928e-09 6.294822e-10 2.470643e-10
#> [29,] 6.362047e-09 2.616418e-09 1.064022e-09 4.285425e-10 1.711563e-10
#> [30,] 3.886444e-09 1.636877e-09 6.804276e-10 2.796374e-10 1.137849e-10
#> [31,] 2.276397e-09 9.825421e-10 4.178082e-10 1.753607e-10 0.000000e+00
#> [32,] 1.283095e-09 5.677139e-10 2.470643e-10 1.059622e-10 0.000000e+00
#> [33,] 6.983586e-10 3.167477e-10 1.410978e-10 0.000000e+00 0.000000e+00
#> [34,] 3.682169e-10 1.711563e-10 0.000000e+00 0.000000e+00 0.000000e+00
#> [35,] 1.886363e-10 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [36,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [37,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [38,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [39,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [40,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [41,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>              [,41]
#>  [1,] 0.000000e+00
#>  [2,] 0.000000e+00
#>  [3,] 0.000000e+00
#>  [4,] 0.000000e+00
#>  [5,] 0.000000e+00
#>  [6,] 0.000000e+00
#>  [7,] 0.000000e+00
#>  [8,] 0.000000e+00
#>  [9,] 0.000000e+00
#> [10,] 0.000000e+00
#> [11,] 0.000000e+00
#> [12,] 0.000000e+00
#> [13,] 0.000000e+00
#> [14,] 0.000000e+00
#> [15,] 1.313010e-10
#> [16,] 1.711563e-10
#> [17,] 2.131567e-10
#> [18,] 2.532449e-10
#> [19,] 2.866764e-10
#> [20,] 3.089320e-10
#> [21,] 3.167477e-10
#> [22,] 3.089320e-10
#> [23,] 2.866764e-10
#> [24,] 2.532449e-10
#> [25,] 2.131567e-10
#> [26,] 1.711563e-10
#> [27,] 1.313010e-10
#> [28,] 0.000000e+00
#> [29,] 0.000000e+00
#> [30,] 0.000000e+00
#> [31,] 0.000000e+00
#> [32,] 0.000000e+00
#> [33,] 0.000000e+00
#> [34,] 0.000000e+00
#> [35,] 0.000000e+00
#> [36,] 0.000000e+00
#> [37,] 0.000000e+00
#> [38,] 0.000000e+00
#> [39,] 0.000000e+00
#> [40,] 0.000000e+00
#> [41,] 0.000000e+00